requirements.txt
flask
python-dotenv
requests
pytz
